requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-bcrypt
flask-wtf
flask-mail
pymysql
python-dotenv
reportlab
openpyxl
requests
pytest
pytest-flask